feat(dapp): support analytics event in boost embed
Creates yet another wrapper component. Moved a bunch of the logic from
`RegistrySection` out into `RegistryWrapper`, in order to cleanly use
dispatch and context in `RegistrySection` to instantiate the fire
analytics event early enough that the boost embed can share that logic
